requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.35
supabase==2.7.4
upstash-redis==1.3.0
httpx==0.27.2
twilio==9.3.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
boto3==1.35.23
duckdb==1.1.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.14.0
pytest==8.3.3
pytest-asyncio==0.24.0
coverage==7.6.1
ruff==0.6.8
mypy==1.11.2
celery==5.4.0
flower==2.0.1
gTTS==2.4.0
